FMWK-788 Improve bandwidth limiter (#315)

+++ cmd/internal/models/common.go
@@ -33,5 +33,5 @@ type Common struct {
 
 	// Nice is mapped to config.Bandwidth
 	// Is set in MiB then converted to bytes.
-	Nice int
+	Nice int64
 }
